MILAB-6013 batch-processing: refactor file layout, docs, and errors
Split batch code into smaller, focused files and cleaned up the public
surface:

- Extract processColumnBatch + shared processColumn helpers
  (buildOutputTrace, createProcessColumnResult, column-entry
  normalization) out of index.lib.tengo into process-column-common and
  process-column-batch libs.
- Extract orchestrator template helpers (classifyPColumnData,
  groupColumnData, etc.) into process-pcolumn-batch.lib.tengo so the
  template stays under 500 lines.
- Re-export the three column-entry helpers from util.lib.tengo with
  @deprecated notices so existing imports keep working.
- Replace if-panic chains with ll.assert; use sets.fromSlice/hasElement
  for enum checks and slices.map for trivial transforms; delegate xsvType
  validation to pUtil.validateXsvType.
- Rewrite every exported function's doc comment in pt/index.lib.tengo
  style (summary + @param/@returns); add file-level overviews listing
  exports.
- Rewrite assert/panic messages to state what happened and what the
  caller should do (value shown, rule, action).
- Split the 1697-line proc_batch.test.ts into five scenario-focused test
  files with a shared proc_batch_common.ts fixture module.
- Add opts.cacheInputs plumbing through the batch path (wf.cacheInputs);
  test harness now defaults to times.second so repeated runs don't reuse
  stale pt caches on the backend.
